Moving business logic from highlighted to titleVisible
Previously the logic of displaying the title was dependent on the
highlighted state. This was suboptimal, because it meant that whenever
the textfield was selected, the title was shown (seen as a blink).

With this change I’ve separated the displaying of the title to be
driven by the setTitleVisible callback instead.